Fix UB in add_*_mconstraint
NLopt's nlopt_add_*_mconstraint functions take a pointer to an array
of tolerances, one for each constraint; they copy that array.

rust-nlopt mistakenly passed a pointer to a single f64.  As a result
NLopt would make out-of-bounds array accesses.  Often the values read
would be negative, giving NLOPT_INVALID_ARGS.

There are two possible ways to fix this: 1. simply fix the rust-nlopt
API to take an appropriate slice; 2. make the existing functions copy
the single provided toleranc value and provide 2 new functions for
when the tolerances are not all the same.

I have chosen (1) because it is less work and because the declared
version of this crate is 0.x, so this would not break semver.

Two more arguments that could be made in support of this decision:

Anyone calling these functions is currently experiencing UB,
including (in my experience) random invalid argument errors.  So
probably few people are using these entrypoints.

The new API corresponds to the NLopt API.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>